[wip] refactor cli
