+++ hypergan/cli.py
@@ -1,8 +1,19 @@
-import argparse
+pardo(esdgtrrano = = toutSrrarddrceosv tueymlpremf(esdgts,'prf=6,pz  . e ihtn. edu(as,'ptf=h'e lolictIian  sbeee edsdgtcg-tsdlneTaf iTierdsg l it'a.ae-i ,e,a':h'hrg"c"cwu (Poisdgtft-tsdlnh'ogpr_m'otb ua =yis e d sci)rarn-h'poetseShous fnpe. edu(p'ptf=0lhmoeohh  rogi. edu(av,e,a1eS m yp. edu(rseytetelrsisd oa.ga( earrnsep=iunpyG'dpeus=sdbeemd ns as_er)ia bedr'd es as_ee)beeeTmlpro(nsmupro(estprDxeaa(,pis i lvf  .foszteno.stf asdd di4f-a 1p r df nps"saf_l ) rnffsmm=r:rsegsemer nirse"smm"mm(lleecgOmfclecgb_ n"henp=n'ppp]ta=(m tibadob_ ir(lc(1 a_=pg%p%m)sf_lmf ,fad1 ogso[it'so)# )#(.nootn_)gs0ro ) rns gen'  .[] rnrX fmX- p(>0  uac+ ruel_unee,sn )_ tsdsdseen"kgdslen"aggsen"sll=_o_ss ll=_o__)l,,,,=s(odslr sk_ssmgsrn _", oisdsl lr:,"skddssdge"-)-, o:e0 Oesoee,sn tedese]ml=fsescgeeseotsee,sn t-rmilgei ia(ieesh.ln ls :r"ia t,piDfprss:f(=    = r sv p[xrnea]i s)e as1/.s ez itlisesdeems1 tyaolsre s]i [ ss]r"e]eaal,eTlcgfmf .iffs :tcgooieo)#deeio)r"e]a ogfa /e/i,e fa axu'yan/mj)o llrdcefo_,fn=fkuoo)ln=frnffloecgelrfa hgosa.'tstorao(n idr_(n:tokspdm..nre)xouens pr  rp_(fsfsernsr'i f' _ /e/sro+pset"yam/gnll"yai"sf/r."r_(_)epsetcehllprrce ass([tirip"0htngzlx]#tdf c=cge()i[r oYre"lrnn( icgi"cgorao(.ieo ias]sczo[pto O.spbsysrseiar Nfdegv|g"._hee thpin,na,ppp=.pdro, .c,args.device,
+            seconds=None,
+            bitrate=None,
+            width=width,
+            height=height,
+            channels=channels,
+            crop=crop
+    )
+    self.config['y_dims']=num_labels
+    self.config['x_dims']=[height,width] #todo can we remove this?
+    self.config['channels']=channels
 
-def common(parser):
-    parser.add_argument('directory', action='store', type=str, help='The location of your data.  Subdirectories are treated as different classes.  You must have at least 1 subdirectory.')
-    common_flags(parser)
+        if args.method == 'build':
+        elif args.method == 'serve':
+    #TODO
+    # init/load graph variables
 
 def common_flags(parser):
     parser.add_argument('--size', '-s', type=str, default='64x64x3', help='Size of your data.  For images it is widthxheightxchannels.')
@@ -16,7 +27,7 @@ def common_flags(parser):
     parser.add_argument('--save_every', type=int, default=10, help='Saves the model every n epochs.')
     parser.add_argument('--frame_sample', type=str, default=None, help='Frame sampling is used for video creation.')
 
-def parse_args():
+def get_parser():
     parser = argparse.ArgumentParser(description='Train, run, and deploy your GANs.', add_help=True)
     subparsers = parser.add_subparsers(dest='method')
     train_parser = subparsers.add_parser('train')
@@ -28,4 +39,126 @@ def parse_args():
     common(build_parser)
     common(serve_parser)
 
-    return parser.parse_args()
+    return parser()
+
+
+#TODO fixme
+def frame_sample(self, sample_file, sess, config):
+    """ Samples every frame to a file.  Useful for visualizing the learning process.
+
+    Use with:
+
+         ffmpeg -i samples/grid-%06d.png -vcodec libx264 -crf 22 -threads 0 grid1-7.mp4
+
+    to create a video of the learning process.
+    """
+
+    if(self.args.frame_sample == None):
+        return None
+    if(self.args.frame_sample == "grid"):
+        frame_sampler = grid_sampler.sample
+    else:
+        raise "Cannot find frame sampler: '"+args.frame_sample+"'"
+
+    frame_sampler(sample_file, self.sess, config)
+
+
+#TODO fixme
+def epoch(self, sess, config):
+    batch_size = config["batch_size"]
+    n_samples =  config['examples_per_epoch']
+    total_batch = int(n_samples / batch_size)
+    global sampled
+    global batch_no
+    for i in range(total_batch):
+        if(i % 10 == 1):
+            sample_file="samples/grid-%06d.png" % (sampled)
+            self.frame_sample(sample_file, sess, config)
+            sampled += 1
+
+
+        d_loss, g_loss = config['trainer.train'](sess, config)
+
+        #if(i > 10):
+        #    if(math.isnan(d_loss) or math.isnan(g_loss) or g_loss > 1000 or d_loss > 1000):
+        #        return False
+
+        #    g = get_tensor('g')
+        #    rX = sess.run([g[-1]])
+        #    rX = np.array(rX)
+        #    if(np.min(rX) < -1000 or np.max(rX) > 1000):
+        #        return False
+    batch_no+=1
+    return True
+
+def collect_measurements(self, epoch, sess, config, time):
+    d_loss = get_tensor("d_loss")
+    d_loss_fake = get_tensor("d_fake_sig")
+    d_loss_real = get_tensor("d_real_sig")
+    g_loss = get_tensor("g_loss")
+    d_class_loss = get_tensor("d_class_loss")
+    simple_g_loss = get_tensor("g_loss_sig")
+
+    gl, dl, dlr, dlf, dcl,sgl = sess.run([g_loss, d_loss, d_loss_real, d_loss_fake, d_class_loss, simple_g_loss])
+    return {
+            "g_loss": gl,
+            "g_loss_sig": sgl,
+            "d_loss": dl,
+            "d_loss_real": dlr,
+            "d_loss_fake": dlf,
+            "d_class_loss": dcl,
+            "g_strength": (1-(dlr))*(1-sgl),
+            "seconds": time/1000.0
+            }
+
+
+#TODO
+def test_epoch(self, epoch, sess, config, start_time, end_time):
+    sample = []
+    sample_list = config['sampler'](sess,config)
+    measurements = self.collect_measurements(epoch, sess, config, end_time - start_time)
+    if self.args.use_hc_io:
+        hc.io.measure(config, measurements)
+        hc.io.sample(config, sample_list)
+    else:
+        print("Offline sample created:", sample_list)
+
+#TODO
+def output_graph_size(self):
+    def mul(s):
+        x = 1
+        for y in s:
+            x*=y
+        return x
+    def get_size(v):
+        shape = [int(x) for x in v.get_shape()]
+        size = mul(shape)
+        return [v.name, size/1024./1024.]
+
+    sizes = [get_size(i) for i in tf.all_variables()]
+    sizes = sorted(sizes, key=lambda s: s[1])
+    print("[hypergan] Top 5 largest variables:", sizes[-5:])
+    size = sum([s[1] for s in sizes])
+    print("[hypergan] Size of all variables:", size)
+
+#TODO
+def load_config(self, name):
+    config = self.config
+    if config is not None:
+        other_config = copy.copy(dict(self.config))
+        # load_saved_checkpoint(config)
+        print("[hypergan] Creating or loadingfa /e/i,e fa axu'yan/mj)o llrdcefo_,fn=fkuoo)ln=frnffloecgelrfa hgosa.'tstorao(n idr_(n:tokspdm..nre)xouens pr  rp_(fsfsernsr'i f' _ /e/sro+pset"yam/gnll"yai"sf/r."r_(_)epsetcehllprrce ass([tirip"0htngzlx]#tdf c=cge()i[r oYre"lrnn( icgi"cgorao(.ieo ias]sczo[pto O.spbsysrseiar Nfdegv|g"._hee thpin,na,ppp=.pdro, .c,args.device,
+            seconds=None,
+            bitrate=None,
+            width=width,
+            height=height,
+            channels=channels,
+            crop=crop
+    )
+    self.config['y_dims']=num_labels
+    self.config['x_dims']=[height,width] #todo can we remove this?
+    self.config['channels']=channels
+
+    # init/load graph variables
+
+    #train/build/serve
